Don't try to change enabled status of current user
The currently logged-in user can't change their own enabled status, however on clicking Save in the user edit dialog the changeUserEnabledStatus method is still called.
Because of this, on server side a ForbiddenException is thrown and the dialog window does not close.
With this commit a check is added not to call the method.